% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{assign_colors}
\alias{assign_colors}
\title{Assign Colors Based on Tertiles}
\usage{
assign_colors(data, color_matrix)
}
\arguments{
\item{data}{A data frame containing tertile columns}

\item{color_matrix}{A 3x3 matrix of colors}
}
\value{
A data frame with added color column
}
\description{
Assign Colors Based on Tertiles
}
