% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calculate_tertile_breaks}
\alias{calculate_tertile_breaks}
\title{Calculate Tertile Breaks}
\usage{
calculate_tertile_breaks(data, variable)
}
\arguments{
\item{data}{A data frame}

\item{variable}{Character string naming the variable}
}
\value{
A numeric vector of break points
}
\description{
Calculate Tertile Breaks
}
