% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calculate_tertiles}
\alias{calculate_tertiles}
\title{Calculate Tertiles for Bivariate Variables}
\usage{
calculate_tertiles(data, var_1, var_2)
}
\arguments{
\item{data}{A data frame containing the variables to analyze}

\item{var_1}{Character string naming the first variable}

\item{var_2}{Character string naming the second variable}
}
\value{
A data frame with added tertile columns
}
\description{
Calculate Tertiles for Bivariate Variables
}
