% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_bivariate_map}
\alias{create_bivariate_map}
\title{Create Bivariate Choropleth Map}
\usage{
create_bivariate_map(
  data,
  var_1,
  var_2,
  color_matrix = NULL,
  custom_labels = NULL
)
}
\arguments{
\item{data}{A spatial data frame}

\item{var_1}{Character string naming first variable}

\item{var_2}{Character string naming second variable}

\item{color_matrix}{Optional custom color matrix}

\item{custom_labels}{Optional vector of custom HTML labels for tooltips}
}
\value{
A leaflet map object
}
\description{
Create Bivariate Choropleth Map
}
