% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_default_color_matrix}
\alias{create_default_color_matrix}
\title{Create Default Color Matrix for Bivariate Choropleth}
\usage{
create_default_color_matrix()
}
\value{
A matrix of color hex codes
}
\description{
Creates a 3x3 color matrix with default colors for bivariate choropleth maps
}
