% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_legend_html}
\alias{generate_legend_html}
\title{Generate Legend HTML}
\usage{
generate_legend_html(var_1, var_2, var_1_breaks, var_2_breaks, color_matrix)
}
\arguments{
\item{var_1}{Character string naming first variable}

\item{var_2}{Character string naming second variable}

\item{var_1_breaks}{Numeric vector of breaks for first variable}

\item{var_2_breaks}{Numeric vector of breaks for second variable}

\item{color_matrix}{Matrix of colors}
}
\value{
HTML string for legend
}
\description{
Generate Legend HTML
}
